refactor: extract confirmation and buttons into separate component
